requirements.txt
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.5
scikit-learn==0.23.2
sklearn==0.0
streamlit==0.72.0
Pillow==8.0.1
